# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthExpiredParams": {
        "additionalProperties": false,
        "description": "401 `invalid_api_key`; takes no parameters.",
        "properties": {},
        "title": "AuthExpiredParams",
        "type": "object"
      },
...
      "FaultProfileCreate": {
        "additionalProperties": false,
        "description": "A new fault profile. Omitted `params` are the scenario's defaults.",
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthExpiredParams"
              },
              {
                "$ref": "#/components/schemas/ScopeDeniedParams"
              },
              {
                "$ref": "#/components/schemas/RateLimitedParams"
              },
              {
                "$ref": "#/components/schemas/UnsupportedParameterParams"
              },
              {
                "$ref": "#/components/schemas/Provider5xxParams"
              },
              {
                "$ref": "#/components/schemas/MalformedJsonParams"
              },
              {
                "$ref": "#/components/schemas/TruncatedStreamParams"
              },
              {
                "$ref": "#/components/schemas/SlowResponseParams"
              },
              {
                "$ref": "#/components/schemas/TimeoutParams"
              }
            ],
            "default": {},
            "title": "Params"
          },
          "probability": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability",
            "type": "number"
          },
          "scenario": {
            "$ref": "#/components/schemas/FaultScenario"
          }
        },
        "required": [
          "name",
          "scenario"
        ],
        "title": "FaultProfileCreate",
        "type": "object"
      },
      "FaultProfileOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "attached_key_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Attached Key Ids",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthExpiredParams"
              },
              {
                "$ref": "#/components/schemas/ScopeDeniedParams"
              },
              {
                "$ref": "#/components/schemas/RateLimitedParams"
              },
              {
                "$ref": "#/components/schemas/UnsupportedParameterParams"
              },
              {
                "$ref": "#/components/schemas/Provider5xxParams"
              },
              {
                "$ref": "#/components/schemas/MalformedJsonParams"
              },
              {
                "$ref": "#/components/schemas/TruncatedStreamParams"
              },
              {
                "$ref": "#/components/schemas/SlowResponseParams"
              },
              {
                "$ref": "#/components/schemas/TimeoutParams"
              }
            ],
            "title": "Params"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          },
          "scenario": {
            "$ref": "#/components/schemas/FaultScenario"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "scenario",
          "params",
          "probability",
          "enabled",
          "expires_at",
          "active",
          "attached_key_ids",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "FaultProfileOut",
        "type": "object"
      },
      "FaultProfileUpdate": {
        "additionalProperties": false,
        "description": "A partial edit: only the fields sent change.\n\n`params` must be sent with `scenario` (they are only meaningful together); a `scenario`\nsent without `params` resets the params to that scenario's defaults. `expires_at` is cleared\nby sending `null`; every other field must have a value when it is sent. `changes()` tells an\nabsent field from one sent as `null`.",
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthExpiredParams"
              },
              {
                "$ref": "#/components/schemas/ScopeDeniedParams"
              },
              {
                "$ref": "#/components/schemas/RateLimitedParams"
              },
              {
                "$ref": "#/components/schemas/UnsupportedParameterParams"
              },
              {
                "$ref": "#/components/schemas/Provider5xxParams"
              },
              {
                "$ref": "#/components/schemas/MalformedJsonParams"
              },
              {
                "$ref": "#/components/schemas/TruncatedStreamParams"
              },
              {
                "$ref": "#/components/schemas/SlowResponseParams"
              },
              {
                "$ref": "#/components/schemas/TimeoutParams"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "probability": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaultScenario"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "FaultProfileUpdate",
        "type": "object"
      },
      "FaultScenario": {
        "description": "A failure the gateway can inject (the Integration Lab); each has its own params.",
        "enum": [
          "auth_expired",
          "scope_denied",
          "rate_limited",
          "unsupported_parameter",
          "provider_5xx",
          "malformed_json",
          "truncated_stream",
          "slow_response",
          "timeout"
        ],
        "title": "FaultScenario",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MalformedJsonParams": {
        "additionalProperties": false,
        "description": "A 200 whose body is cut to `keep_fraction` of its length; non-streaming calls only.",
        "properties": {
          "keep_fraction": {
            "default": 0.6,
            "maximum": 0.9,
            "minimum": 0.1,
            "title": "Keep Fraction",
            "type": "number"
          }
        },
        "title": "MalformedJsonParams",
        "type": "object"
      },
...
      "Provider5xxParams": {
        "additionalProperties": false,
        "description": "The provider's error at `status`.",
        "properties": {
          "status": {
            "default": 500,
            "enum": [
              500,
              502,
              503,
              529
            ],
            "title": "Status",
            "type": "integer"
          }
        },
        "title": "Provider5xxParams",
        "type": "object"
      },
...
      "RateLimitedParams": {
        "additionalProperties": false,
        "description": "429 with `Retry-After: retry_after_s`.",
        "properties": {
          "retry_after_s": {
            "default": 2,
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Retry After S",
            "type": "integer"
          }
        },
        "title": "RateLimitedParams",
        "type": "object"
      },
...
      "ScopeDeniedParams": {
        "additionalProperties": false,
        "description": "403 `insufficient_permissions`; takes no parameters.",
        "properties": {},
        "title": "ScopeDeniedParams",
        "type": "object"
      },
...
      "SlowResponseParams": {
        "additionalProperties": false,
        "description": "The first byte of the response delayed by `delay_ms`.",
        "properties": {
          "delay_ms": {
            "default": 5000,
            "maximum": 60000.0,
            "minimum": 100.0,
            "title": "Delay Ms",
            "type": "integer"
          }
        },
        "title": "SlowResponseParams",
        "type": "object"
      },
...
      "TimeoutParams": {
        "additionalProperties": false,
        "description": "The call held for `hold_ms`, or what is left of the route's budget, then timed out.",
        "properties": {
          "hold_ms": {
            "default": 30000,
            "maximum": 120000.0,
            "minimum": 1000.0,
            "title": "Hold Ms",
            "type": "integer"
          }
        },
        "title": "TimeoutParams",
        "type": "object"
      },
...
      "TruncatedStreamParams": {
        "additionalProperties": false,
        "description": "A stream closed after `after_chunks` frames, without its end marker.",
        "properties": {
          "after_chunks": {
            "default": 3,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "After Chunks",
            "type": "integer"
          }
        },
        "title": "TruncatedStreamParams",
        "type": "object"
      },
...
      "UnsupportedParameterParams": {
        "additionalProperties": false,
        "description": "400 naming `param` as unsupported by the model.",
        "properties": {
          "param": {
            "default": "temperature",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z_][a-z0-9_.]*$",
            "title": "Param",
            "type": "string"
          }
        },
        "title": "UnsupportedParameterParams",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "created_at",
          "email_verified"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Spanlight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/gateway/fault-profiles": {
      "get": {
        "description": "By name. `active` says whether a profile can fire now: enabled and not expired.",
        "operationId": "list_profiles_api_v1_projects__project_id__gateway_fault_profiles_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FaultProfileOut"
                  },
                  "title": "Response List Profiles Api V1 Projects  Project Id  Gateway Fault Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List fault profiles",
        "tags": [
          "gateway"
        ]
      },
      "post": {
        "description": "Omitted `params` are the scenario's defaults; a param that does not belong to the scenario\nis `422` on `params.<name>`. `409 FAULT_PROFILE_NAME_TAKEN` for a name already used in the\nproject, and `422` on `expires_at` for a time that is not in the future.",
        "operationId": "create_profile_api_v1_projects__project_id__gateway_fault_profiles_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a fault profile",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/projects/{project_id}/gateway/fault-profiles/{profile_id}": {
      "delete": {
        "description": "Delete the profile. Keys that ran it are detached and run no fault.",
        "operationId": "delete_profile_api_v1_projects__project_id__gateway_fault_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a fault profile",
        "tags": [
          "gateway"
        ]
      },
      "patch": {
        "description": "Change the fields sent; send `null` for `expires_at` to remove the expiry.\n\n`params` must be sent with `scenario`; a `scenario` sent alone resets the params to that\nscenario's defaults.",
        "operationId": "update_profile_api_v1_projects__project_id__gateway_fault_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit a fault profile",
        "tags": [
          "gateway"
        ]
      }
    },
...
  }
}